extract media player from media main
